fix app structure
