+++ src/pages/DetailsPage.jsx
@@ -1,4 +1,4 @@
-import Details from '../components/DetailsPage/Details';
+import Details from '../components/Details';
 
 export default function DetailsPage() {
   return <Details />;
